- test with Float insted Double in Collection Lists

+++ server/src/main/java/com/mpsdevelopment/biopotential/server/cmp/machine/Pattern.java
@@ -9,7 +9,7 @@ public interface Pattern {
     String getKind();
     String getName();
     String getDescription();
-    List<Double> getPcmData();
+    List<Float> getPcmData();
     List<ChunkSummary> getSummary();
     String getFileName();
     void setName(String str);
