+++ server/src/main/java/com/mpsdevelopment/biopotential/server/cmp/machine/PcmDataSummary.java
@@ -7,9 +7,9 @@ import com.mpsdevelopment.biopotential.server.cmp.analyzer.ChunkSummary;
 public class PcmDataSummary {
 
 	private final List<ChunkSummary> summary;
-	private final List<Double> pcmData;
+	private final List<Float> pcmData;
 
-	public PcmDataSummary(List<Double> pcmData, List<ChunkSummary> summary) {
+	public PcmDataSummary(List<Float> pcmData, List<ChunkSummary> summary) {
 		this.pcmData = pcmData;
 		this.summary = summary;
 	}
@@ -18,7 +18,7 @@ public class PcmDataSummary {
 		return summary;
 	}
 
-	public List<Double> getPcmData() {
+	public List<Float> getPcmData() {
 		return pcmData;
 	}
