+++ server/src/main/java/com/mpsdevelopment/biopotential/server/cmp/machine/strains/WAWPattern.java
@@ -39,7 +39,7 @@ public class WAWPattern implements Pattern {
 
     }
 
-    public List<Double> getPcmData() {
+    public List<Float> getPcmData() {
         return null;
     }
